Split database_functions.ts into 3 Services
